requirements.txt
alembic==1.7.3
asgiref==3.4.1
Django==2.2.24
django-filter==2.4.0
djangorestframework==3.12.4
greenlet==1.1.1
Mako==1.1.5
MarkupSafe==2.0.1
psycopg2==2.8.6
pytz==2021.1
SQLAlchemy==1.4.25
sqlparse==0.4.2
